Serverless/requirements.txt
boto3
requests
stomp.py